refactored student restrictions
