+++ sources/packages/web/src/services/http/StudentApi.ts
@@ -8,7 +8,6 @@ import {
   StudentProfileAPIOutDTO,
   StudentUploadFileAPIOutDTO,
   UpdateStudentAPIInDTO,
-  StudentRestrictionAPIOutDTO,
   AESTStudentProfileAPIOutDTO,
   SINValidationsAPIOutDTO,
   CreateSINValidationAPIInDTO,
@@ -74,16 +73,6 @@ export class StudentApi extends HttpBaseClient {
    */
   async applyForPDStatus(): Promise<void> {
     await this.patchCall(this.addClientRoot("students/apply-pd-status"), null);
-  }
-
-  /**
-   * API client to call the student restriction rest API.
-   * @returns student restriction(wrapped by promise).
-   */
-  async getStudentRestriction(): Promise<StudentRestrictionAPIOutDTO[]> {
-    return this.getCallTyped<StudentRestrictionAPIOutDTO[]>(
-      this.addClientRoot("students/restriction"),
-    );
   }
 
   /**
